Added background task settings

+++ Viddy/ViewModel/SettingsViewModel.cs
@@ -1,7 +1,10 @@
-﻿using Windows.UI.Xaml;
-using GalaSoft.MvvmLight.Command;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Windows.UI.Xaml;
 using JetBrains.Annotations;
 using ThemeManagerRt;
+using Viddy.Core.Extensions;
 using Viddy.Model;
 using Viddy.Services;
 
@@ -12,20 +15,30 @@ namespace Viddy.ViewModel
         private readonly ISettingsService _settingsService;
         private readonly ITaskService _taskService;
 
+        private bool _ignoreChanges;
+
         public SettingsViewModel(ISettingsService settingsService, ITaskService taskService)
         {
             _settingsService = settingsService;
             _taskService = taskService;
+
+            _ignoreChanges = true;
+            UpdateFrequencies = Enum.GetValues(typeof (UpdateFrequency)).ToList<UpdateFrequency>();
+            NotificationFrequency = UpdateFrequencies.First(x => x == _settingsService.NotificationFrequency.GetFrequency());
+            _ignoreChanges = false;
         }
 
         public bool IsLightTheme { get; set; }
         public bool IsLocationOn { get; set; }
         public bool CheckForNotificationsInBackground { get; set; }
-        public int NotificationCheckFrequencyInMinutes { get; set; }
+        public UpdateFrequency NotificationFrequency { get; set; }
+        public List<UpdateFrequency> UpdateFrequencies { get; set; }
 
         [UsedImplicitly]
         private void OnCheckForNotificationsInBackgroundChanged()
         {
+            if (_ignoreChanges) return;
+            
             if (CheckForNotificationsInBackground)
             {
                 _taskService.CreateService();
@@ -39,12 +52,14 @@ namespace Viddy.ViewModel
         }
 
         [UsedImplicitly]
-        private void OnNotificationCheckFrequencyInMinutesChanged()
+        private void OnNotificationFrequencyChanged()
         {
+            if (_ignoreChanges) return; 
+            
             _taskService.RemoveService();
-            _taskService.CreateService(NotificationCheckFrequencyInMinutes);
+            _taskService.CreateService(NotificationFrequency.GetMinutes());
 
-            _settingsService.NotificationCheckFrequencyInMinutes = NotificationCheckFrequencyInMinutes;
+            _settingsService.NotificationFrequency = NotificationFrequency.GetMinutes();
         }
 
         [UsedImplicitly]
